+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/controller/TestController.java
@@ -1,4 +1,4 @@
-package com.avdo.spring_app.controller;
+package com.avdo.spring.app.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
