+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package com.avdo.spring.app.repository;
+
+import com.avdo.spring.app.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
