--- backend/spring-app/spring-app/src/main/java/com/avdo/spring_app/repository/ProfileRepository.java
@@ -1,7 +1,0 @@
-package com.avdo.spring_app.repository;
-
-import com.avdo.spring_app.model.Profile;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProfileRepository extends JpaRepository<Profile, Long> {
-}
