+++ backend/spring-app/spring-app/src/test/java/com/avdo/spring/app/SpringAppApplicationTests.java
@@ -1,4 +1,4 @@
-package com.avdo.spring_app;
+package com.avdo.spring.app;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
